+++ classes/prefix_WPgutenberg/blocks/vimeo/attributes.js
@@ -1,0 +1,28 @@
+const attributes = {
+  videoID: {
+    type: 'string',
+    default: '',
+  },
+  videoAutoPlay: {
+    type: 'boolean',
+    default: true,
+  },
+  videoBackgroud: {
+    type: 'boolean',
+    default: false,
+  },
+  videoLoop: {
+    type: 'boolean',
+    default: false,
+  },
+  videoDimentionX: {
+    type: 'string',
+    default: '4',
+  },
+  videoDimentionY: {
+    type: 'string',
+    default: '3',
+  }
+}
+
+export default attributes;